requirement.txt
aiohttp == 3.7.4.POST0
aiopg == 1.2.1
aioredis == 1.3.1
databases == 0.4.3
locust == 1.5.1
passlib == 1.7.4
bcrypt == 3.2.0
asyncpg == 0.22.0
PyJWT == 2.1.0
python-multipart == 0.0.5
flake8 == 3.9.2